+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 3.0.2
 
 * Added a clear-all button
+* Implemented more Vaadin-Mixins for better customization
 * Fixed the required indicator
   * The indicator of the "available items"-ComboBox is now only present when the wrapping ChipComboBox is empty
   * The indicator value is now correctly returned by ``isRequiredIndicatorVisible``